pyproject.toml
[project]
name = "basoka-psn-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "requests>=2.32.4",
    "ruff>=0.12.7",
]
